refactor(auth): move AuthHandle/Principal to runtime/auth.rs (backend-agnostic)
Mirror the runtime/images.rs precedent exactly: relocate `Principal`,
the `pub(crate) trait AuthBackend`, and `pub struct AuthHandle` from
`rest/auth.rs` into `runtime/auth.rs` (always-compiled, alongside
`runtime/images.rs`). `BoxliteRuntime::auth()` and the `auth_backend`
field are ungated; field type becomes
`Option<Arc<dyn crate::runtime::auth::AuthBackend>>`.

`RestRuntime` remains the only `AuthBackend` impl, with the `impl`
block still co-located in `rest/runtime.rs` (where the private
`ApiClient` field is in scope). `rest/auth.rs` deleted;
`rest::types::Principal` removed (the deserialization test moved
with it into `runtime/auth.rs::tests`).

This places identity at the same architectural layer as image ops
rather than living rest-feature-gated. Local runtimes still return
`BoxliteError::Unsupported` from `auth()`, mirroring `images()`
returning `Unsupported` for REST runtimes.

Public surface delta vs the previous commit on this branch:
`AuthHandle` and `Principal` are no longer `#[cfg(feature = "rest")]`-
gated -- they are always part of the crate root (consistent with
`ImageHandle` being public/ungated).

Verified: cargo check (non-rest --lib + boxlite-cli), clippy
`--all-targets -D warnings` (boxlite --no-default-features,
boxlite --features rest, boxlite-cli), 6 auth integration tests,
4 serve auth unit tests, moved `test_principal_deserialization`.
