+++ src/boxlite/src/rest/mod.rs
@@ -5,7 +5,6 @@
 //!
 //! Enabled with the `rest` feature flag.
 
-pub(crate) mod auth;
 pub(crate) mod client;
 pub mod credential;
 pub(crate) mod error;
